runtime/pyproject.toml
[project]
name = "benny"
version = "1.0.0"
description = "Deterministic Graph Workflow Platform with Multi-Model AI Orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "litellm>=1.60.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "chromadb>=0.5.0",
    "openlineage-python>=1.27.0",
    "arize-phoenix>=5.0.0",
    "python-multipart>=0.0.20",
    "pypdf>=5.0.0",
    "pathspec>=0.12.0",
    "tiktoken>=0.7.0",
    "rich>=13.0.0",
    "neo4j>=5.25.0",
    "networkx>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-repeat>=0.9.0",
    "ruff>=0.8.0",
]
mcp = [
    "mcp>=1.0",
]

[project.scripts]
benny = "benny_cli:main"
benny-release = "benny.ops.release_tool:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
